+++ templates/bootstrap/app.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'theme/app_theme.dart';
 {{#if go_router}}import 'router.dart';
 {{/if}}{{#unless go_router}}import '../features/home/presentation/pages/home_page.dart';
 {{/unless}}{{#if l10n}}import '../l10n/app_localizations.dart';
@@ -10,11 +11,8 @@ class App extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp{{#if go_router}}.router{{/if}}(
       title: '{{project_name}}',
-      theme: ThemeData(
-        useMaterial3: {{#if material3}}true{{/if}}{{#unless material3}}false{{/unless}},
-{{#if material3}}        colorScheme: ColorScheme.fromSeed(seedColor: Colors.blue),
-{{/if}}{{#unless material3}}        primarySwatch: Colors.blue,
-{{/unless}}      ),
+      theme: AppTheme.light,
+      darkTheme: AppTheme.dark,
 {{#if l10n}}      localizationsDelegates: AppLocalizations.localizationsDelegates,
       supportedLocales: AppLocalizations.supportedLocales,
 {{/if}}{{#if go_router}}      routerConfig: appRouter,